Flashlight sfx cooldown fixed.
